dom/tests/mochitest/general/test_focus_legend_noparent.html

<html>
<head>
<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<script type="text/javascript">

SimpleTest.waitForExplicitFinish();

function runTest()
{
    window.focus();
    var g = document.createElementNS("http://www.w3.org/1999/xhtml", "legend");
    document.body.appendChild(g);
    setTimeout(g.focus.bind(g), 0);
    setTimeout(done, 10);
}

function done()
{
  ok(document.hasFocus(), "document is still focused");
  is(document.activeElement, document.body, "document has no focused element")
  SimpleTest.finish();
}

SimpleTest.waitForFocus(runTest);
</script>
</head>

<p id="display"></p>
<div id="content" style="display: none">
</div>

<body onblur="dump('blurred window!\n')"></body>
</html>
